+++ tomcat/src/main/java/org/apache/coyote/controller/FrontController.java
@@ -8,10 +8,10 @@ import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.Map;
 import org.apache.coyote.http11.HttpHeader;
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
 import org.apache.coyote.http11.HttpStatusCode;
 import org.apache.coyote.http11.MimeType;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.util.FileExtension;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
