+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -7,6 +7,8 @@ import java.io.OutputStream;
 import java.net.Socket;
 import org.apache.coyote.Processor;
 import org.apache.coyote.controller.FrontController;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
