+++ tomcat/src/main/java/org/apache/coyote/http11/request/RequestLine.java
@@ -1,7 +1,8 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 import java.util.HashMap;
 import java.util.Map;
+import org.apache.coyote.http11.HttpMethod;
 
 public class RequestLine {
